requirements.txt
torch
transformers
evaluate
datasets
tqdm
hydra-core
scikit-learn
pandas
openai